--- src/routes/InquiryRouter.ts
@@ -1,22 +1,0 @@
-import { Router } from "express";
-import { body } from "express-validator/check";
-
-import { InquiryController } from "../controllers";
-
-const router = Router();
-
-router.post(
-  "/user/inquiry",
-  [
-    body("userId").notEmpty(),
-    body("email").notEmpty(),
-    body("inquiryCategory").notEmpty(),
-    body("orderNum").notEmpty(),
-    body("title").notEmpty(),
-    body("content").notEmpty(),
-    body("isSubscribed").notEmpty(),
-  ],
-  InquiryController.createInquiry,
-);
-
-export default router;
