+++ src/routes/inquiry.ts
@@ -1,9 +1,23 @@
 import { Router } from "express";
+import { body } from "express-validator/check";
 
 import { inquiryController } from "../controllers";
 
 const router: Router = Router();
 
+router.post(
+  "/user",
+  [
+    body("userId").notEmpty(),
+    body("email").notEmpty(),
+    body("inquiryCategory").notEmpty(),
+    body("orderNum").notEmpty(),
+    body("title").notEmpty(),
+    body("content").notEmpty(),
+    body("isSubscribed").notEmpty(),
+  ],
+  inquiryController.createInquiry,
+);
 router.delete("/:inquiryId", inquiryController.deleteInquiry);
 
 export default router;
